+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # MiniReel Player 2.0
 
 * *[v1.5.1-rc1]*
+  * Update the Vzaar player to be consistent with the behavior of newer players
   * Update the Vine player to be consistent with the behavior of newer players
 
 ## v1.5.0 (November 20, 2015)